Made requested changes refer below for details
Commented the old distance_to_anomaly_gdal as there were too many functions for the user. Added an exception in the tests for gdal_array as the gdal_array is available only on the Windows system. Further more updates in the notebook are made as the old distance_to_anomaly_gdal is commented and hence is removed here.